completed code for sneak peak and started footer
started the code for the sneak peak section on the home page. added all the images, as well as corresponding text. added a button beneath the section taking you to flights. started the footer, i have added the logo, as well as the home, contact, flight which all take you back, as well as added the social media links. styles to come.